requirements.txt
tree_sitter
ast
javalangcd 
chromadb
transformers
faiss-cpy